C89: fix decl-after-statement + enum-range regressions breaking linux-c89 CI
Four C89 violations introduced by earlier audit-fix bundles on this branch
(never CI-validated, since local/* branches don't trigger CI) that fail the
strict `make C89_BUILD=1` linux-c89 job (retroarch.yml):

* runloop.h:129 — RUNLOOP_FLAG_IS_INITED = (1u << 31) exceeds INT_MAX, so the
  enum constant is out of C89 int-range (-Werror=pedantic). Cast to
  (int)(1u << 31): keeps the audit's unsigned-shift UB avoidance while the
  constant stays representable as int. (regressed by d6bbeeab55)
* runloop.c — SET_MESSAGE env-cb: the NULL-guard split msg's declaration from
  its initialisation, stranding the HAVE_GFX_WIDGETS p_dispwidget declaration
  after code. Hoist it above the guard. (regressed by 46855caa12)
* verbosity.c — rarch_log_file_init: the buffer [0]='\0' inits were placed
  before the static/bool declarations. Move the inits below. (regressed by
  ab7f4d0f6e)
* menu_setting.c — setting_append_input_bind_list: the !list guard was placed
  before the defaults/binds_group_label declarations. Move it below them.
  (regressed by bd9fc145d8)

Found by the new local-CI.sh linux-c89 mirror. Verified green: c89 job
(0 repo-source errors across C89_BUILD + DEBUG builds) and the linux-i686
container job (full build + link in CI's xenial-gcc9 image).

+++ runloop.h
@@ -126,7 +126,9 @@ enum runloop_flags
    RUNLOOP_FLAG_IDLE                              = (1u << 28),
    RUNLOOP_FLAG_FOCUSED                           = (1u << 29),
    RUNLOOP_FLAG_FORCE_NONBLOCK                    = (1u << 30),
-   RUNLOOP_FLAG_IS_INITED                         = (1u << 31)
+   /* (int) cast keeps this bit-31 flag within the C89 enum int-range while
+    * the unsigned shift still avoids signed-overflow UB (audit S1+S10). */
+   RUNLOOP_FLAG_IS_INITED                         = (int)(1u << 31)
 };
 
 /* Contains the current retro_fastforwarding_override
